requirements.txt
aiohttp~=3.9.0
Brotli~=1.1.0
yarl~=1.9.3
python-dotenv~=1.0.0
requests~=2.31.0

